Move test fixtures into their own package.
