+++ src/test/java/cruise/umple/umpr/core/consistent/ConsistentsTest.java
@@ -20,8 +20,8 @@ import com.jayway.jsonassert.JsonAsserter;
 import cruise.umple.umpr.core.DiagramType;
 import cruise.umple.umpr.core.ImportFSM;
 import cruise.umple.umpr.core.Repository;
-import cruise.umple.umpr.core.repositories.TestRepository;
-import cruise.umple.umpr.core.util.MockModule;
+import cruise.umple.umpr.core.fixtures.MockModule;
+import cruise.umple.umpr.core.fixtures.TestRepository;
 
 @Guice(modules={MockModule.class})
 @Test
