+++ src/test/java/cruise/umple/umpr/core/fixtures/MockModule.java
@@ -1,4 +1,4 @@
-package cruise.umple.umpr.core.util;
+package cruise.umple.umpr.core.fixtures;
 
 import com.google.inject.AbstractModule;
 
@@ -6,7 +6,6 @@ import cruise.umple.umpr.core.DocumentFactory;
 import cruise.umple.umpr.core.RealDocumentFactory;
 import cruise.umple.umpr.core.consistent.ConsistentsModule;
 import cruise.umple.umpr.core.entities.EntityModule;
-import cruise.umple.umpr.core.repositories.TestRepositoryModule;
 
 /**
  * Test module to load files locally instead of from web pages, allowing for
